fix(DuckDB path logic for dbt profiles): Presererve full path when pulling from dbt Profile
We do some logic to make the Path for a DuckDB database file a relative path when filled out in the manual form. The path defined in a DuckDB dbt profile will be a full path though. This fixes the logic so dbt Profile paths are unchanged while manual form paths are relative.