feat(cli)!: Refactor to use Cobra
This PR doesn't functionally change anything about the internal workings of the app, but greatly alters the structure of the app. It now uses Cobra for a more proper CLI-like experience (version, help, etc work). This is laying the ground work for saving configs for easier reuse of previous runs.